streamHits() returns a stream of hits from scroll search
This is the core feature of the library. Given the url of a standard
search request and a hash of elasticsearch search parameters, this
function will continuously stream the hits from the search api into a
single readable stream.

Consumers only need to understand request bodies that can be sent to the
standard search api, and then all of the scrolling logic is handled
internally.

If a scroll duration is set in the query string of the given url, it
will be honored, otherwise the default is a 30s scroll duration.

For example:

```js
const url = 'http://localhost:9200/logstash-*/_search';
const params = { query: { term: { response: 200 } } };

streamHits(url, params).pipe(someOtherStream);
```