Bubble any errors up to the main hit stream
Whenever any streaming or http errors occur, the error details are now
bubbled all the way up and emitted on the primary hit stream. Whenever
an error is encountered, streaming is halted immediately.

This required a pretty significant refactor to better separate concerns
throughout the library so bubbling up errors didn't result in
significant coupling across the various components. As a nice side
effect, the individual modules are much easier to reason about and less
brittle.